LayoutTests/fast/forms/week/week-stepup-stepdown-from-renderer.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Check stepping-up and -down for <input> from renderer. No cases of empty initial values for type=week.');

var input = document.createElement('input');
var invalidStateErr = '"Error: INVALID_STATE_ERR: DOM Exception 11"';

function sendKey(keyName) {
    var event = document.createEvent('KeyboardEvent');
    event.initKeyboardEvent('keydown', true, true, document.defaultView, keyName);
    input.dispatchEvent(event);
}

function setInputAttributes(min, max, step, value) {
    input.min = min;
    input.max = max;
    input.step = step;
    input.value = value;
}

function stepUp(value, step, max, optionalStepCount) {
    setInputAttributes(null, max, step, value);
    if (typeof optionalStepCount != "undefined")
        if (optionalStepCount < 0)
            for (var i = 0; i < -optionalStepCount; i++)
                sendKey('Down');
        else
            for (var i = 0; i < optionalStepCount; i++)
                sendKey('Up');
    else
        sendKey('Up');
    return input.value;
}

function stepDown(value, step, min, optionalStepCount) {
    setInputAttributes(min, null, step, value);
    if (typeof optionalStepCount != "undefined")
        if (optionalStepCount < 0)
            for (var i = 0; i < -optionalStepCount; i++)
                sendKey('Up');
        else
            for (var i = 0; i < optionalStepCount; i++)
                sendKey('Down');
    else
        sendKey('Down');
    return input.value;
}

input.type = 'week';
debug('Function arguments are (value, step, {min or max}, [stepCount]).');
debug('Normal cases');
shouldBe('stepUp("2010-W02", null, null)', '"2010-W03"');
shouldBe('stepDown("2010-W02", null, null)', '"2010-W01"');
shouldBe('stepUp("2010-W02", null, null, 10)', '"2010-W12"');
shouldBe('stepDown("2010-W02", null, null, 11)', '"2009-W44"');
shouldBe('stepUp("1970-W01", "4", null, 2)', '"1970-W09"');
shouldBe('stepDown("1970-W01", "4", null, 3)', '"1969-W41"');
debug('Step=any');
shouldBe('stepUp("2010-W02", "any", null)', '"2010-W03"');
shouldBe('stepDown("2010-W02", "any", null)', '"2010-W01"');
debug('Overflow/underflow');
shouldBe('stepUp("2010-W02", "3.40282346e+38", null)', '"275760-W37"');
shouldBe('stepDown("2010-W02", "3.40282346e+38", null)', '"1970-W01"');
shouldBe('stepUp("2010-W02", "1", "2010-W02")', '"2010-W02"');
shouldBe('stepDown("2010-W02", "1", "2010-W02")', '"2010-W02"');
debug('stepDown()/stepUp() for stepMismatch values');
shouldBe('stepDown("2010-W02", "2", "2009-W52")', '"2010-W01"');
shouldBe('stepUp("1970-W02", "4", "")', '"1970-W05"');

debug('');
var successfullyParsed = true;
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
