LayoutTests/fast/forms/week/week-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for  from renderer. No cases of empty initial values for type=week.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
PASS stepUp("2010-W02", null, null) is "2010-W03"
PASS stepDown("2010-W02", null, null) is "2010-W01"
PASS stepUp("2010-W02", null, null, 10) is "2010-W12"
PASS stepDown("2010-W02", null, null, 11) is "2009-W44"
PASS stepUp("1970-W01", "4", null, 2) is "1970-W09"
PASS stepDown("1970-W01", "4", null, 3) is "1969-W41"
Step=any
PASS stepUp("2010-W02", "any", null) is "2010-W03"
PASS stepDown("2010-W02", "any", null) is "2010-W01"
Overflow/underflow
PASS stepUp("2010-W02", "3.40282346e+38", null) is "275760-W37"
PASS stepDown("2010-W02", "3.40282346e+38", null) is "1970-W01"
PASS stepUp("2010-W02", "1", "2010-W02") is "2010-W02"
PASS stepDown("2010-W02", "1", "2010-W02") is "2010-W02"
stepDown()/stepUp() for stepMismatch values
PASS stepDown("2010-W02", "2", "2009-W52") is "2010-W01"
PASS stepUp("1970-W02", "4", "") is "1970-W05"

PASS successfullyParsed is true

TEST COMPLETE
